Rework the endpoint acquire/release code to avoid oops when the last closer if the BH:
Thanks to the previous commit, the application close request
doesn't need to wait for refcount to be null before releasing
stuff. It is sure that nobody will be able to acquire the
endpoint anymore, we can return to the application immediately.

The last refcount user (either the application close request
or the BH receiving a packet) takes care of releasing all the
endpoint resources. Since vfree cannot be called from BH context,
it just move the endpoint away on a cleanup list that the cleanup
kernel thread will take care of regularly (as we already do for
pull handles).

kref are used instead of atomic counters to handle this nicely.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@914 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356