Revert commits 913 and 918: * Stop dereferencing endpoint->iface everywhere in the driver.   Just get the iface when acquire the resources at the beginning.   This will make it much easier to detach endpoint. * Cache the interface mac address in our iface structure   to reduce the need to dereference its link to the kernel   interface structure (iface->eth_ifp).
This is useless now that I understand how to hold/release interfaces
properly even when the unregister notifier is called. There's no need
to have all endpoints closed when returning from the notifier, we just
need to make sure they will be closed soon, and thus the reference on
the iface will be released soon too (and the driver will wait for in
rtnl_unlock).

We still detach the endpoint in the actual close, but we keep
endpoint->iface valid until the last endpoint reference is released
so that current processing can finish safely. They will release
the endpoint then, and the last one will release the iface. And the
last one will release the device and wakeup rtnl_unlock.

The next step is to use kref from endpoint to iface and it should
finally be safe to detach ifaces.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@924 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356