Change the pull handle magic into the combination of: * a 24bits generation incremented with each new handle * 8bits for the endpoint index (xor'ed with a constant magic number)
Check this whole new magic instead of only the 8bits endpoint index
when receiving a pull reply or a nack mcp. This way, we can detect 
obsolete pull replies that could have matched a new handle with same
idr index (the idr is likely to always allocate the same small indexes).


The 24bits generation should theoritically be included in the wire 
handle. But Linux does not provide any API to do that easily. So we
keep the idr in the full 32bits wire handle and use 24bits of the 
magic to make the wire handle unique.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1020 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356