Use the iface peer field instead of a new peer structure to link a local peer into the peer table/hash. And maintain it across iface attach/detach and peer clearing/adding.
This way, setting the peer table immediately renames ifaces according to their
peer name. And omx_hostname renames the peer in the table automatically.

When an interface is detached, the peer is removed from the table/hash without
adding a "non-local" peer to replace it. So the peer entry is just lost.
But we don't really care about this. It is still possible to re-add the peer
table if needed.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1025 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356