Make MPICH1-MX build and Open MPI configure/build without requiring any change in their code and/or configure script.
We add a omx__mx_compat.c file in the lib containing non-inlined
MX wrappers so that the final Open-MX lib contains ABI wrappers
required by configure checks (such as mx_finalize in Open MPI)
or direct symbol references (such as &mx_test in MPICH1-MX).

This is not ABI compatibility, but it helps a lot already.

Only mx_test and mx_finalize are ABI-wrapped so far, more to
come later after some reorganization of the API wrappers.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1219 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356