When the application requests an explicit disconnect, free the partner structure and remove it from the endpoint table, and assume the corresponding endpoint address isn't valid anymore.
It prevents from getting "reconnect, needs to cleanup"
messages later when the disconnect wasn't decided internally.

MX doesn't do this, but its behavior isn't really documented,
and I think it's a sane design.


Add omx_disconnect() to omx_pingpong when running in slave
mode and going back to waiting for a new client.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1320 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356