omx__error_with_req receives some request status code, not a return code. That's kind of a mess. Cast status codes into return codes for now.
We should either (1) use a different error_handler for request status code,
or (2) just merge status code and return code into a single enum.

(2) might be better as of today since nobody uses the Open-MX API yet,
and requiring to rebuild MX applications because of an ABI breakage is
not that bad as this point of the development.

Opinions are welcome...



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1416 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356