The current regcache can reuse the same window for multiple larges at the same time as long as there is only one for send.
Make this disabled if only OMX_RCACHE is passed (to match MX behavior)
and enable this "parallel regcache mode" with OMX_PRCACHE is set.

We can't support multipl sender at the same time because the sender
"request" is identified by the rdmawin id. We'll need a 32bit opaque
identifier in the rndv/notify fields to improve this.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1596 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356