When completing a pull handle whose memory copies have been offloaded to a dma engine, do not busy poll in the BH until their completion arrives.
It's better to only check for their completion once, and
schedule a workstruct if needed to do the actual busy poll
and completion outside of the BH.

This requires a HAL for workstruct manipulation.

This seems to fix some possible lockup under heavy load,
possibly caused by the BH polling too much and preventing
other I/O AT stuff from running. Might well be a I/O AT
hardware bug according to the maintainer, and this would
be a lucky workaround...



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1598 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356