Check skb length and linearity before dereferencing early fields: * only dereference the packet type if it is in the linear skb header,   revert to skb_copy_bits if not * compute the header length and use it to linearize the whole Open-MX   header if needed (instead of using the maximal header size)
It fixes the support for wrongly-small skbuff (with a debug check in case
it's smaller than ETH_ZLEN), and highly-non-linear skbuffs.

No bug has ever been observed in a released tarball since all headers are
smaller than ETH_ZLEN, which guarantees that skbuff will always be large
enough for any packet type.
However, if a packet type ever gets a header >44B (i.e. 60B with the
ethernet header), we can receive some headers that are smaller than the
maximal header size, causing the previous code to drop the packet.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1718 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356