Fix a long-standing bug in the unexp handler FINISHED case: If FINISHED is given for a RNDV message, we need to notify its completion to the sender first, we can't just drop it entirely.
Since this is a very rare case, we abort if we fail to allocate
the temporary notify request.

We could have made FINISHED an invalid application behavior for
RNDV (the application can post a recv to discard it manually).
But the application doesn't know for sure whether a message
is RNDV or not. So let's just try to support this since it's
not that hard.

Should be backported in MX.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2228 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356