Fix management of pull block descriptors in case of many packet loss at the end of a large message: * make sure we always shift the first non-completed block   up to the first desc; * make sure we always prepare as much desc as possible instead   of stopping when we fail to actually send a pull request.
This fixes a BUG_ON(!handle->block_desc[0].frames_missing_bitmap)
in the timeout handler routine. It was caused by completed blocks
not being shifted away when we were requesting the last blocks
of a message and we lost a packet in block N-3 and N-1.

Thanks a lot to Duwayne D. Mulhall for the detailed bug report
and analysis.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2242 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356