omx_init: Add udev support
Create device nodes only if udev is not running or if they
don't exist yet.
chgrp/mod device nodes only if we created them above, or
udev is not running, or if the udev rules file is absent.

This way, if udev is running and the rules files was installed
by omx_local_install, udev should take care of everything and
the admin will be able to tweak the rules without having omx_init
override it.

Configure-given defaults are only enforced if udev is absent
or if we can't find the udev rules (and thus assume udev may
not do eveything as expected).

Add several messages to clarify all this during omx_init.
Also print ls -l of each device node for easier check.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2312 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356