Support 16kB sendq/recvq slots
In case somebody enforces --with-medium-frag-max=8952
to fully use MTU=9000 (the default is 8kB frags to avoid
wasting 7kB out of 16kB slots but it may not matter much).

For the record, myri10ge actually prefers 4kB frag,
we may do --with-medium-frag-max=4096 by default in
some other NICs behave the same.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2408 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356