Work-around deadlocks that could occur when multiple endpoints are open in the same process and they are connecting in random order.
If a single-threaded process is sleeping in a blocking call,
other endpoints are never polled, so incoming connection
requests are never processed. This leads to easy deadlocks
when process A connects to rail #0 of B while process B
connects to rail#1 of A.

Add OMX_CONNECT_POLLALL=1 to force mx_connect() to poll other
endpoints as well. It is enough for now since most MPI layers
don't use any other blocking calls. Another solution would be
to add a progression thread.

If open several endpoints without OMX_CONNECT_POLLALL=1 in the
environment, a verbose message is displayed.

Thanks to Dave Love for reporting the problem.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2637 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356