regcache: Implement mx__regcache_clean()
Add omx_regcache_clean() and use it in mx__regcache_clean().
This is needed because OpenMPI 1.3 assumes that it works and thus
enables regcache by default with MX_RCACHE=2.

So Open-MX still does not enable regcache by default because it
does not intercept malloc properly, but it lets OpenMPI take care
of regcache management by just providing mx__regcache_clean().

Thanks to Matthew Korzeniowski for testing.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2733 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356