Fix a bogus assertion failure that may occur when out-of-order early fragments are received:
We had a special case putting early at the beginning of the partner
queue if its seqnum is lower than all other early packets. And we
assume all other cases would never put the early at the beginning.

This was wrong since we also have to put at the beginning if it's
a medium frag with same seqnum than the first early in the queue
and with a lower medium *frag* seqnum.

To simplifies all this, we just remove the special cases about
inserting at the beginning or at the end of the queue. The latter
is handled by the first iteration in the main queue traversal loop.
The former is just a final fallback if we found no other valid
position in the queue during the traversal loop.

This caused an assertion failure on SciNet/U-Toronto with 256 and
512 processes on 64 nodes.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2784 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356