Fix a potential memory corruption of the destination buffer address when receiving expected messages:
When a receive request doesn't match immediately, we copy the given
segments into a newly allocated request. But we didn't update the segs
pointer when a single segment was used, so it would point to the old
(possibly freed/reused) segment descriptor.

Always use omx_clone_segments so that segs properly points to single
when needed.

It's not clear whether this may have caused actual problems on site.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2846 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356