Work-around issues with the kernel build system trying to create files in the build directory during make install.
Kbuild may fail during make install as a different user since it tries to create
some files in .tmp_versions even when there is nothing to do. This is especially
a problem when doing make install as root over NFS without root squash.
Ignore these errors when the current user is not the one that ran configure.
So make install as root does not fail, and we still do not ignore build failures
as the normal user.

By the way, update TODO, and add some documentation in Kbuild.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2926 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356