Workaround overzealous gcc 4.4 warnings enabled in 2.6.33 on x86_32.
If copy_from_user (or other memcpy-like functions) cannot check
the target buffer size, they complain with:
  In function 'copy_from_user',
    inlined from 'omx_shared_send_tiny' at driver/linux/omx_shared.c:257:
  asm/uaccess_32.h:212: warning: call to 'copy_from_user_overflow' declared with attribute warning: copy_from_user() buffer size is not provably correct

We already check the given length, but only in the caller, not in the
shared-memory code. So we need to recheck the length again here so as
to help the compiler understand what's going on.

Moreover, for some reason, checking directly 'hdr->length' isn't enough,
we need an additional intermediate value 'length'.


If this kind of mess happens again, we'll seriously think about disabling
this kernel warning entirely...



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2931 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356