Fixed a build system issue regarding to -m flag
the GCC -m flag is not supported on all architectures (ex: IA64). Yet,
the build system used to always specify the -m flag, according to the
target architecture, even if multilib was not activated. This caused
building to fail on archs not supporting -m (error: unrecognized flag)

Fixed by specifying the -m flag only when --enable-multilib is given.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@2979 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356