When thread-safety is enabled (default), define USE_LOCKS in dlmalloc so as to protect it from concurrent thread accesses. It fixes the case where multiple threads access multiple endpoints in the same process at the same time (probably never occurs in any real world application).
The overhead for non-multithreaded programs is negligible (additional
nop-locking in glibc per request submission)
.
The overhead for multihthreaded programs (as soon as libphread is loaded)
is a couple microseconds per request submission. We might improve this
by switching to another allocation library (not a good idea after -rc1,
we'll see later).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@3006 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356