Add a new memory allocation API
This new API coexists with the existing one : omx_malloc(),omx_calloc(),omx_free()
but takes an endpoint as first parameter :
    omx_malloc_ep(ep, ...)
    omx_calloc_ep(ep, ...)
    omx_free_ep(ep, ...)

This new API will be used, in place of the old one, everywhere the
endpoint function parameter is already locked.
This could allow to use in the future a lockless per-endpoint allocator.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@3094 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356