When we receive a nack lib for a connect request, the seqnum is always 0, it's not the connect_seqnum (since the MX MCP doesn't know about the connect opaque data internals). Just complete the first connection request for this peer, they are all equivalent anyway.
This doesn't change anything for program that issued a single connect
request to each peer, since the connect_seqnum would be 0 too.
But if multiple connect requests were sent to an invalid peer, the
non-first ones would not be completed properly since the nack seqnum
(0) and request connect_seqnum (>0) are different.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@3096 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356