Fix in-order matching when many unexpected messages arrive out-of-order
We must make sure that the endpoint unexpected queues remain in-order
so that matching is performed in order for both expected and unexpected
messages.
So we can't dequeue requests and requeue them later when processing
intermediate fragments of unexpected medium messages.

Thanks to Guillermo Lopez Taboada for the report and test case.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@3098 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356