Rework exp/unexp queue offsets to fix and simplify their management
Use indexes instead of offsets, and keep this index on the whole uint32_t
range so that we can distinguish when they are identical or only identical
modulo the length of the queue.

Once the index is properly assigned and checked, derive the slot offset
and address as well as the event id from it.

By the way, all these offset are also better renamed:
* next_ -> nextfree_
* last_free_ -> nextreleased_ (off by one)
* next_reserved_ -> nextreserved_

By the way again, update the cancelling of reserved slots to the
event-id based model.