Update the number of available expected event slots to the new delivery mechanism
Up to BATCH_NR-1 event slots may not have been freed yet, so we
can't let the kernel use these event slots immediatly.

When an expected event is processed, avail_exp_event is increased
immediately, but the slot may actually be released later in another
call to omx__progress().

We could optimize this by looking at how many slots actually need
to be released. But it would likely require that the number of
expected slots is updated in the main event processing loop and
not in event-specific callback anymore. It's not very important
anyway, we still have hundreds of available expected slots.