Change the wait ioctl to pass event indexes instead of offsets
Now that both lib and driver use omx_eventq_index_t to find the next event slot+id,
directly compare these index instead of an offset in the wait ioctl.
Add some build-time assertions to make sure that these omx_eventq_index_t can be
stored in the ioctl and endpoint descriptors (it's easier to keep uint32_t there
to ensure that the padding is correct).