Start a locking policy rework related to events
omx_ioctl_wait_event() now uses RCU-flavor functions for
insertions/deletions in the event waiter list.

As a result, the omx_event_waiter object being inserted in and removed
from the list is now dynamically allocated.
Its freeing is delayed in a RCU callback (a rcu_head struct has been
added to the omx_event_waiter struct).

This patch, which is the first of two ones, implements a RCU approach on
the writer side of the event waiter list.