fixed: kernel allocation in a spinlock section
The dynamic allocation of the event waiter introduced in
a recent commit was done in a spinlock section. This is strictly
forbidden as the kmalloc() may sleep.
Fixed by moving the allocation outside of the spinlock section.

Moreover, this allocation is now checked (as well as on the library side
in omx_test.c)