Replace the current list implementation in the library
The current list implementation used in the library is borrowed
from the Linux kernel sources and is released under GPL, leading to
a contamination inside the library.
To avoid this, this patch drops the current list implementation and
provides a new one written from scratch which exposes a quite similar
interface (yet slightly different to better fit our needs).
For this occasion, the code of this implementation moves from common/ to
libopen-mx/.