Finally make --with-linux-release the main option for changing the target kernel
make kernelrelease isn't reliable (broken again on Debian's 2.6.38),
so we can't use its output to decide where to install the module.
Otherwise omx_init would fail to find/load the module later if
uname -r is different from make kernelrelease.

Install the module in kernel release path specified by the user,
which means uname -r by default, or what was given with --with-linux-release.
Only warn if make kernelrelease reports something different.

Require --with-linux-release before --with-linux{,-build} because
we cannot use uname -r when compiling on a non-running kernel
(and make kernelrelease isn't reliable).
It's not a problem anyway because passing --with-linux-release is
usually much easier than passing the entire kernel header path
in --with-linux{,-build}, and the former will usually be enough.