Merging Xen-specific changes to Open-MX (backend).
This is a merged patchset for the integration of Open-MX into the Xen split driver model. You can find more info for the
initial deployment in README.xen.

There's a lot of duplication in the sense that common/unchanged open-mx files exist both in the backend and frontend
directories under the driver subdir. Before the aplha release we plan on stripping everything down to a minimum, to
avoid uneccessary duplication.

The building procedure is exactly the same, however installing is untested and should be tailored to the modular aspect
of this tree. We have to carefully rename the modules that corespond to the backend and the frontend, thus, installing
is not yet supported.

The system is experimental -- there are a lot of bugs, memory leaks that render the system unstable after the initial
insmod (both in the driver domain and the guest).

Locking is buggy, leading to corrupt data if multiple processes try to send/receive data from a single guest, or if
multiple guests try to communicate with the network.

All in all, it is a first step towards an alpha release, uploaded in order to get comments/feedback.

This is the backend part.