Added timers to the frontend (and the backend).
enhancing the ones for the backend too. Pretty lame and hacky but we get a first
impression of where the time is spent. The most important part (latency wise) is
the frontend kick to the backend, taking aprox 14us. If we use a hybrid
workqueue/poll approach, this could lead to almost identical latency with the
pci-attached case. Will branch of master to have a closer look.