Hacky patch to poll on the backend side.
We poll for a couple of micros if we still have requests to handle. This gives
us almost identical latency to the pci-attached case ;-). However, this is a
hacky patch. we need to integrate it smoothly with the rest of the requests.

A rough sketch, could be to parse the request on the be_int() handler, and then,
depending on the type of the request (TINY, SMALL, MEDIUM, LARGE), push the
request into the relevant workqueue, or handle it immediately. This way, we
could optimize both small/tiny sends (LATENCY!!!) and large ones (for efficient
pipelining, which we don't have yet).