Refactoring ring message handling.
We need to have two paths optimized for latency and bandwidth. In order to be
able to trasmit data to the network fast enough, we need the backend to be able
to kick the relevant functions immediately. At the same time, we need to be able
to pipeline large requests in order to get the necessary bandwidth.

We refactor the process_message workqueue handler, to be able to choose if we
will handle requests inside the first level handler (be_int), or queue the work
struct to be handled later (process_message_request). This is the first patch to
be able to keep a consistent way of handling requests.