Porting the cookies infrastructure.
We now support cookies for a set of pages (per segment) and not for just one page. Thus, every time we register a segment, we search for a cookie that fits the number of pages. If we haven't got one, we allocate it, and when we deregister this segment, we put it back in the free list. So, we don't need to allocate pages every time we register. We just do that when it is absolutely necessary.