Getting rid of a message in cookies.
...and sanitizing a perfect bug ;-)

In the backend cookies case (the set of pre-allocated pages for accepting grants
in the backend), we ask for a cookie by *_get_cookie and the first time (when
the list is empty, the code throws us to alloc_cookies(). However, if this path
is crossed more than OMX_XEN_MAX_COOKIES we see this as an error and abort.

THIS IS NOT CORRECT. We keep this for debugging purposes only, and add a message
to let us know what's going on, instead of just staring at the kernel trace and
wonder ;-)