Implementing some kind of fe<->be shortcut.
In order to minimize the latency in region reg/dereg, we add this buggy option.
If the region is considered created or destroyed at the backend, then the
frontend gets poked without having to wait for the response to come and handle
it via the common way.

This way we're only delayed by 2us (instead of polling for 10s or 100s).
however, the real bottleneck lies in the grant operations ... still unsolved :S