Adding the dmaengine functionality to the backend.
The backend now supports off-loading memcopies to the DMA engines provided by
the platform. Unfortunately we are not able to improve performance by using
IOATDMAs, however, it seems like a nice feature, we should be able to exploit
afterwards ;) (for instance, when we will have solved the increased latency when
accepting grants)

The two files are almost identical to the original source files. The only change
that added this functionality is the different region pointer we pass on
initialization (xregion vs. region). This is because we stricly follow Open-MX
semantics throughout the implementation; hence a small change for a really
great feature (which is not yet visible, but ;-) anyway).